back to using double instead of float (float only in hit.h and hit.cpp)

+++ include/monte_carlos.hpp
@@ -10,10 +10,10 @@
 #include "arguments_parsing.hpp"
 #include "random_generator.hpp"
 
-float monte_carlo_single(size_t N);
+double monte_carlo_single(size_t N);
 
-float monte_carlo_auto_parallel(size_t N, const Arguments& args);
+double monte_carlo_auto_parallel(size_t N, const Arguments& args);
 
-float monte_carlo_manual_parallel(size_t N, const Arguments& args);
+double monte_carlo_manual_parallel(size_t N, const Arguments& args);
 
-float calculate_volume(size_t N, const Arguments& args);
+double calculate_volume(size_t N, const Arguments& args);
